wallets/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
python-dotenv==1.0.1
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
redis==5.1.1
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

